Cleaned Up Code
I went through the code and removed all the text code.  All the text code that was worth keeping I placed at the end of the files in comments.  All comments where given names.

Quickly added the words to learn to an array and hooked up to the queryUser method.  Will test that now.