Program Generates Array Of Random Numbers
The program now generates an array of random numbers based on the number of words to learn.  There are no duplicates in the array.  Added a general text file to write down random notes.  Added a process file to note the current process.  Added to the toDoList.  Added words to learn to a text file for save keeping.

Updated The To Do List

Cleaned Up Code

Added comments to the code as well as reorganising the flow.